Update sagas to share same code
